Add SCSS support
This implements #79.

This adds support to typeface packages by introducing a `_typeface.scss`
partial, which at present exists to facilitate setting a font-path
variable for a given typeface. This is to enable support of Sass-centric
build systems, such as via sass directly, where font files will be
copied/put into a known location by another process.

The implementation means this is possible:

```
$open-sans-font-path:          "../fonts/open-sans" !default;
@import "typeface-open-sans/typeface";
```

and the resulting CSS is populated with the correct font-path as set.

The naming of the imported partial is arbitrary -- it could just as
easily be `fonts` or something else. Ideally, it would have been
`index.scss` or `_index.scss` to enable a simple `@import
"typeface-open-sans"` from within Sass, but this causes ambiguity
with the existing `index.css` file and Sass won't build if both files
are present (it used to be a warning and is now an error).

The choice to have typeface-specific variable names gives flexibility
over one `$font-path` variable so you can set different paths for
different typefaces (plus being more specific means less chance of a
conflict in existing projects).

The implementation in the code itself adds only a few minor changes to
the build script and templates. There are no changes to the existing
CSS build for full backward compatibility.